+++ packages/core/src/types/EventTypes.ts
@@ -47,6 +47,8 @@ type VoiModifiedEventDetail = {
   invert?: boolean;
   /** Indicates if the 'invert' state has changed from the previous state */
   invertStateChanged?: boolean;
+  /** color map */
+  colormap?: ColormapPublic;
 };
 
 type ColormapModifiedEventDetail = {